ft: clear autocomplete after submission
