+++ apps/hash-frontend/src/pages/shared/top-context-bar/share-dropdown-menu/invite-account-form.tsx
@@ -1,18 +1,55 @@
+import { Autocomplete, Avatar } from "@hashintel/design-system";
 import { AccountGroupId, AccountId } from "@local/hash-subgraph";
-import { Box } from "@mui/material";
-import { FunctionComponent, useCallback, useState } from "react";
+import { autocompleteClasses, Box, outlinedInputClasses } from "@mui/material";
+import { FunctionComponent, useCallback, useMemo, useState } from "react";
 
-import { MinimalOrg, MinimalUser } from "../../../../lib/user-and-org";
+import { useOrgs } from "../../../../components/hooks/use-orgs";
+import { useOrgsWithLinks } from "../../../../components/hooks/use-orgs-with-links";
+import { useUsers } from "../../../../components/hooks/use-users";
+import { useUsersWithLinks } from "../../../../components/hooks/use-users-with-links";
+import {
+  MinimalOrg,
+  MinimalUser,
+  Org,
+  User,
+} from "../../../../lib/user-and-org";
 import { Button } from "../../../../shared/ui";
-import { AccountSelector } from "./account-selector";
+import { getImageUrlFromEntityProperties } from "../../get-image-url-from-properties";
 
 export const InviteAccountForm: FunctionComponent<{
   excludeAccountIds?: (AccountId | AccountGroupId)[];
   onInviteAccount: (account: MinimalOrg | MinimalUser) => void;
 }> = ({ onInviteAccount, excludeAccountIds }) => {
-  const [selectedAccount, setSelectedAccount] = useState<
-    MinimalOrg | MinimalUser
-  >();
+  const [selectedAccount, setSelectedAccount] = useState<User | Org | null>(
+    null,
+  );
+
+  const [search, setSearch] = useState("");
+  const [open, setOpen] = useState(false);
+
+  const { users: minimalUsers } = useUsers();
+  const { orgs: minimalOrgs } = useOrgs();
+
+  const { orgs } = useOrgsWithLinks({
+    orgAccountGroupIds: minimalOrgs?.map((org) => org.accountGroupId),
+  });
+  const { users } = useUsersWithLinks({
+    userAccountIds: minimalUsers?.map((user) => user.accountId),
+  });
+
+  const options = useMemo(
+    () =>
+      [...(users ?? []), ...(orgs ?? [])].filter(
+        (account) =>
+          !excludeAccountIds ||
+          !excludeAccountIds.includes(
+            account.kind === "user"
+              ? account.accountId
+              : account.accountGroupId,
+          ),
+      ),
+    [excludeAccountIds, orgs, users],
+  );
 
   const handleSubmit = useCallback(
     (event: React.FormEvent<HTMLFormElement>) => {
@@ -20,6 +57,8 @@ export const InviteAccountForm: FunctionComponent<{
 
       if (selectedAccount) {
         onInviteAccount(selectedAccount);
+        setSelectedAccount(null);
+        setSearch("");
       }
     },
     [selectedAccount, onInviteAccount],
@@ -32,9 +71,64 @@ export const InviteAccountForm: FunctionComponent<{
       columnGap={0.75}
       onSubmit={handleSubmit}
     >
-      <AccountSelector
-        excludeAccountIds={excludeAccountIds}
-        onSelect={(account) => setSelectedAccount(account)}
+      <Autocomplete<User | Org | null, false, false>
+        inputProps={{
+          endAdornment: null,
+        }}
+        autoFocus={false}
+        options={options}
+        inputPlaceholder="Add a user or organization..."
+        open={open}
+        disableClearable={false}
+        onOpen={() => setOpen(true)}
+        onClose={(_, reason) => {
+          if (reason !== "toggleInput") {
+            setOpen(false);
+          }
+        }}
+        inputValue={search}
+        value={selectedAccount}
+        onInputChange={(_, value) => setSearch(value)}
+        onChange={(_, account) => setSelectedAccount(account)}
+        getOptionLabel={(option) =>
+          option?.kind === "user"
+            ? option.preferredName ?? ""
+            : option?.name ?? ""
+        }
+        renderOption={(props, option) => {
+          if (!option) {
+            return null;
+          }
+          const avatarSrc = option.hasAvatar
+            ? getImageUrlFromEntityProperties(
+                option.hasAvatar.imageEntity.properties,
+              )
+            : undefined;
+          return (
+            <Box component="li" {...props}>
+              <Avatar
+                src={avatarSrc}
+                title={
+                  option.kind === "user" ? option.preferredName : option.name
+                }
+                size={28}
+                sx={{ marginRight: 1 }}
+              />
+              {option.kind === "user" ? option.preferredName : option.name}
+            </Box>
+          );
+        }}
+        height={36}
+        sx={{
+          height: 36,
+          [`.${outlinedInputClasses.root}`]: {
+            height: "unset",
+            paddingY: 0.75,
+            [`.${autocompleteClasses.input}`]: {
+              paddingY: 0,
+            },
+          },
+        }}
       />
       <Button disabled={!selectedAccount} size="xs" type="submit">
         Invite
